+++ Pandora/Paradigm/Inventory.hs
@@ -3,4 +3,4 @@ module Pandora.Paradigm.Inventory (module Exports) where
 import Pandora.Paradigm.Inventory.Optics as Exports
 import Pandora.Paradigm.Inventory.Storage as Exports
 import Pandora.Paradigm.Inventory.State as Exports
-import Pandora.Paradigm.Inventory.Environmental as Exports
+import Pandora.Paradigm.Inventory.Environment as Exports